Issue #1318 - Drop in performance for dragging tracks since March build

+++ org.mwc.debrief.legacy/src/Debrief/Wrappers/Track/ITimeVariableProvider.java
@@ -31,7 +31,7 @@ public interface ITimeVariableProvider
 	 * @param dtg
 	 * @return
 	 */
-	Double getValueAt(HiResDate dtg);
+	long getValueAt(HiResDate dtg);
 
 	/** whether the user wants the errors to be drawn to scale on the spatial plot
 	 *
